+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="v0.0.4"></a>
+### v0.0.4 (2013-12-02)
+
+
+#### Features
+
+* **postMessage:**
+  * receive a message ([3cba0fff](https://github.com/kenhkan/angular-xdmessage/commit/3cba0fff340318d192ad48a57d2940b738ea3b54))
+  * send a message ([5a687d8f](https://github.com/kenhkan/angular-xdmessage/commit/5a687d8ff349b8d155b69e6c5a98815010657550))
+  * add postMessage module on top of xdmessage ([e49de13e](https://github.com/kenhkan/angular-xdmessage/commit/e49de13e6869f6c32eff5b20945f6d2610338a47))
+
 <a name="v0.0.3"></a>
 ### v0.0.3 (2013-12-02)